refactor: separacion de responsabilidades en LoadingContext.tsx
