--- src/contexts/LoadingContext.css
@@ -1,23 +1,0 @@
-body {
-    background-color: var(--ion-color-medium-contrast); /* Fondo basado en el tema */
-}
-  
-.loader-overlay {
-position: fixed;
-top: 0;
-left: 0;
-width: 100%;
-height: 100%;
-background: var(--ion-color-medium-contrast);
-opacity: 0.75;
-display: block;
-justify-content: center;
-align-content: center;
-align-items: center;
-z-index: 1000;
-}
-
-.ball {
-    width: 50%;
-    height: 50%;
-}
